Update to use DISCO-L475VG-IOT01A.
Updated to 5.15.8 of mbed-os
Added support for WiFi on the DISCO-L475VG-IOT01A, by adding the wifi
stack etc.

The bind function provided by ISM43362Interface.cpp is modified to
always return success. This is the simplest way of making the WiFi stack
work with the client, but probably not the best. To do the is originali
file is replaced with an almost exact copy and the original added
to .mbedignore.

Added the ability to use the default setting via the menu.